move some 4xx tests in test_users_4xx.py and add some new tests
